+++ src/blocks/Section.ts
@@ -19,6 +19,7 @@ import { Block, Field } from 'payload/types';
 
 import { heading } from '../fields/heading';
 import { richTextFields } from '../fields/link';
+import useAppendEmptyParagraph from '../hooks/useAppendEmptyParagraph';
 import { deepMerge } from '../utils/deepMerge';
 
 import ButtonLink from './ButtonLink';
@@ -26,13 +27,18 @@ import Gallery from './Gallery';
 import ImageStack from './ImageStack';
 import MessagesMarquee from './MessagesMarquee';
 import PortfolioCards from './PortfolioCards';
+import Quotes from './Quotes';
+import Stepper from './Stepper';
 
-const singleColumnBlocks = [ButtonLink, PortfolioCards, Gallery, MessagesMarquee];
+const singleColumnBlocks = [ButtonLink, PortfolioCards, Gallery, MessagesMarquee, Quotes, Stepper];
 const multiColumnBlocks = [ButtonLink, ImageStack];
 
 const richTextField = (columns: '1' | '2'): Field => ({
   name: 'content',
   type: 'richText',
+  hooks: {
+    beforeValidate: [useAppendEmptyParagraph],
+  },
   editor: lexicalEditor({
     features: () => [
       HeadingFeature({ enabledHeadingSizes: ['h1', 'h2', 'h3'] }),